Rename module containing errors to `errors`.
