+++ client-js/src/client/errors.js
